requirements.txt
-i https://pypi.org/simple
absl-py==2.1.0; python_version >= '3.7'
annotated-types==0.7.0; python_version >= '3.8'
astunparse==1.6.3
blis==1.3.0; python_version < '3.14' and python_version >= '3.6'
build==1.2.2.post1; python_version >= '3.8'
catalogue==2.0.10; python_version >= '3.6'
certifi==2024.12.14; python_version >= '3.6'
charset-normalizer==3.4.1; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
cloudpathlib==0.21.0; python_version >= '3.9'
confection==0.1.5; python_version >= '3.6'
contourpy==1.3.1; python_version >= '3.10'
cycler==0.12.1; python_version >= '3.8'
cymem==2.0.11
filelock==3.16.1; python_version >= '3.8'
flatbuffers==24.12.23
fonttools==4.55.3; python_version >= '3.8'
fsspec==2024.12.0; python_version >= '3.8'
gast==0.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-pasta==0.2.0
grpcio==1.69.0; python_version >= '3.8'
h5py==3.12.1; python_version >= '3.9'
idna==3.10; python_version >= '3.6'
imutils==0.5.4
jinja2==3.1.5; python_version >= '3.7'
joblib==1.4.2; python_version >= '3.8'
kagglehub==0.3.6; python_version >= '3.9'
keras==3.8.0; python_version >= '3.9'
kiwisolver==1.4.8; python_version >= '3.10'
langcodes==3.5.0; python_version >= '3.9'
language-data==1.3.0
libclang==18.1.1
marisa-trie==1.2.1; python_version >= '3.7'
markdown==3.7; python_version >= '3.8'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
matplotlib==3.10.0; python_version >= '3.10'
mdurl==0.1.2; python_version >= '3.7'
ml-dtypes==0.4.1; python_version >= '3.9'
mpmath==1.3.0
murmurhash==1.0.12; python_version >= '3.6'
namex==0.0.8
networkx==3.4.2; python_version >= '3.10'
numpy==2.0.2; python_version >= '3.9'
nvidia-cublas-cu12==12.4.5.8; python_version >= '3'
nvidia-cuda-cupti-cu12==12.4.127; python_version >= '3'
nvidia-cuda-nvrtc-cu12==12.4.127; python_version >= '3'
nvidia-cuda-runtime-cu12==12.4.127; python_version >= '3'
nvidia-cudnn-cu12==9.1.0.70; python_version >= '3'
nvidia-cufft-cu12==11.2.1.3; python_version >= '3'
nvidia-curand-cu12==10.3.5.147; python_version >= '3'
nvidia-cusolver-cu12==11.6.1.9; python_version >= '3'
nvidia-cusparse-cu12==12.3.1.170; python_version >= '3'
nvidia-nccl-cu12==2.21.5; python_version >= '3'
nvidia-nvjitlink-cu12==12.4.127; python_version >= '3'
nvidia-nvtx-cu12==12.4.127; python_version >= '3'
opencv-python==4.11.0.86; python_version >= '3.6'
opt-einsum==3.4.0; python_version >= '3.8'
optree==0.14.0; python_version >= '3.8'
packaging==24.2; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
pillow==11.1.0; python_version >= '3.9'
preshed==3.0.9; python_version >= '3.6'
protobuf==5.29.3; python_version >= '3.8'
pybind11==2.13.6; python_version >= '3.7'
pydantic==2.11.3; python_version >= '3.9'
pydantic-core==2.33.1; python_version >= '3.9'
pygments==2.19.1; python_version >= '3.8'
pyparsing==3.2.1; python_version >= '3.9'
pyproject-hooks==1.2.0; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.2
requests==2.32.3; python_version >= '3.8'
rich==13.9.4; python_full_version >= '3.8.0'
scikit-learn==1.6.1; python_version >= '3.9'
scipy==1.15.1; python_version >= '3.10'
seaborn==0.13.2; python_version >= '3.8'
sentencepiece==0.2.0
setuptools==80.0.0; python_version >= '3.9'
shellingham==1.5.4; python_version >= '3.7'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
smart-open==7.1.0; python_version >= '3.7' and python_version < '4.0'
spacy==3.8.5; python_version < '3.13' and python_version >= '3.9'
spacy-legacy==3.0.12; python_version >= '3.6'
spacy-loggers==1.0.5; python_version >= '3.6'
srsly==2.5.1; python_version < '3.14' and python_version >= '3.9'
sympy==1.13.1; python_version >= '3.8'
tensorboard==2.18.0; python_version >= '3.9'
tensorboard-data-server==0.7.2; python_version >= '3.7'
tensorflow==2.18.0; python_version >= '3.9'
tensorflow-cpu==2.18.1; python_version >= '3.9'
tensorflow-io-gcs-filesystem==0.37.1; python_version < '3.13' and python_version >= '3.7'
termcolor==2.5.0; python_version >= '3.9'
thinc==8.3.6; python_version < '3.14' and python_version >= '3.9'
threadpoolctl==3.5.0; python_version >= '3.8'
torch==2.5.1; python_full_version >= '3.8.0'
torchinfo==1.8.0; python_version >= '3.7'
torchtext==0.6.0
torchvision==0.20.1; python_version >= '3.8'
tqdm==4.67.1; python_version >= '3.7'
triton==3.1.0
typer==0.15.3; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
typing-inspection==0.4.0; python_version >= '3.9'
tzdata==2024.2; python_version >= '2'
urllib3==2.3.0; python_version >= '3.9'
wasabi==1.1.3; python_version >= '3.6'
weasel==0.4.1; python_version >= '3.7'
werkzeug==3.1.3; python_version >= '3.9'
wheel==0.45.1; python_version >= '3.8'
wrapt==1.17.2; python_version >= '3.8'
